Overwrite the `libcuda.so` symlink if present
As the `libcuda.so` symlink could have been added during a previous
activation (and deactivation may not have occurred yet, like logging
out), the symlink may already be there when we try to create it. In this
case, just go ahead and forcibly overwrite the symlink with a new one.