Drop custom `conda_build_config.yaml`
As this content is already in `conda-forge-pinning` and we are now using
`cuda_compiler_version` to configure the builds of `nvcc`, try dropping
the custom `conda_build_config.yaml` from this recipe. We should be able
to pick up all of the content from `conda-forge-pinning`. After all
`cuda_compiler_version` is zipped to the `docker_image`, which gets
added anyways. Also should avoid some unnecessary duplication of
information.