Skip when `cuda_compiler_version` is `None`
As we are only interested in builds of `nvcc` that line up to actual
CUDA versions, adjust the `skip` to also handle the case when
`cuda_compiler_version` is `None`. This only exists for feedstocks that
build a CPU-only package in one case, which is not relevant here.
Should avoid this case being included in the re-rendering.