MNT: Re-rendered with conda-build 3.18.9, conda-smithy 3.6.0, and conda-forge-pinning 2019.10.11
Now that we rely on `conda-forge-pinning` to determine the versions of
`nvcc` to build and what Docker image to use to perform those builds,
re-render the feedstock to update the CI content accordingly. Everything
winds up being basically the same (with some minor renaming). Should
simplify the maintenance of this feedstock a bit.