Leverage global pinnings to configure `nvcc`
Drop the bulk of `conda_build_config.yaml` in favor of using the values
in the global pinnings. This cuts down on duplication of this content
and thus makes it easier to pull in the latest changes (like recent
Docker image and CDT changes).

Also adapt the recipe slightly so that we only pick out the `nvcc`
configuration. Plus make sure conda-smithy recognizes this is a CUDA
using recipe even though it does not use `compiler("cuda")` in the
build. Do this by adapting the `skip` to check `compiler("cuda")`
matches `nvcc*`.